Try all formats when mounting filesystems to write files
The filesystem to mount is either newly created by Ignition or reused
because the specification matches. That means Ignition knows the
filesystem and can use a single mount system call. However, if we want
to allow reusing existing filesystems with a different format, it would
be better to have a mount behavior like the mount utility has that
tries all filesystems until success.

Try all allowed formats when mounting filesystems to write files. Due
to the specification and the validator existing filesystems that are
not in the hardcoded list are rejected anyway and we don't need to
query more dynamically.