Fix null pointer deref crash when no OEM filesystem label is specified
The check for reusing a filesystem didn't cover the case where no
filesystem label was specified for the OEM partition, causing a crash.

Instead of handling the missing filesystem labels in the check when an
existing OEM partition is checked for reuse, we relax the conditions
for reuse of the OEM partition and will always reuse it unless a
mismatching FS label or UUID got specified by the user (and the user
doesn't force a wipe).