internal/disk/fs: ignore fs format mismatches for the OEM partition
As soon as the OEM partition's filesystem format changes, all users
that have file entries for the OEM partition (e.g., for kernel
parameters), need to also switch to the new format, otherwise Ignition
will bootloop due to the encountered mismatch error. If users have set
the wipeFilesystem option there is no change in behavior and the user
would create the desired filesystem on each Ignition run.

To spare the users the migration due to an internal detail of the OEM
partition, allow the OEM filesystem format to mismatch and just reuse
the existing OEM partition. The user can still enable the
wipeFilesystem to reformat regardless what content the partition has.

Signed-off-by: Mathieu Tortuyaux <mtortuyaux@microsoft.com>